src/requirements.txt
anyio==3.6.1
azure-common==1.1.28
azure-core==1.25.1
azure-keyvault-secrets==4.5.1
azure-mgmt-compute==27.2.0
azure-mgmt-core==1.3.2
azure-mgmt-resource==21.1.0
azure-mgmt-storage==20.1.0
azure-storage-blob==12.13.1
fastapi==0.78.0
pymongo==4.2.0
python-dotenv==0.20.0
python-multipart==0.0.5
typing-extensions==4.2.0
uvicorn==0.18.2
Flask==2.1.2
Jinja2==3.1.2
pydocumentdb==2.3.5